cloud_functions/harvest/timesheet_data_lake/pyproject.toml
[tool.poetry]
name = "timesheet"
version = "0.1.0"
description = ""
authors = ["urquha <andy.urquhart@hotmail.co.uk>"]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
google-cloud-core = "^2.4.1"
google-cloud-bigquery = "^3.25.0"
google-cloud-secret-manager = "^2.20.0"
pyarrow = "^16.1.0"
pandas-gbq = "^0.23.1"
holidays = "^0.52"
data_pipeline_tools = { git = "https://github.com/tpximpact/GCP-Data-Pipelines-DX.git", branch = "main" }
pyforecast = { git = 'https://github.com/urquha/py_forecast', branch = "master" }
pandas = "^2.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
